pyproject.toml
[project]
name = "ltdb-rag-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "beautifulsoup4>=4.14.3",
    "bitsandbytes>=0.49.0",
    "colpali-engine>=0.3.13",
    "einops>=0.8.1",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-postgres>=0.0.16",
    "langgraph>=1.0.5",
    "lxml>=6.0.2",
    "markdownify>=1.2.2",
    "pgvector>=0.3.6",
    "pillow>=12.1.0",
    "playwright>=1.57.0",
    "psycopg2-binary>=2.9.11",
    "pytest-playwright>=0.7.2",
    "pyyaml>=6.0.3",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.1",
    "torch>=2.8.0",
    "transformers>=4.57.3",
]
